requirements.txt
gspread
google-auth
dbt-core
dbt-redshift
astronomer-cosmos
psycopg2-binary
apache-airflow-providers-amazon
apache-airflow-providers-google
pandas
pyarrow
apache-airflow